Ignore supervisor service on the target state
Starting with v3 state endpoint, the supervisor will receive the configuration
for the supervisor service on the target state. This commit allows the
supervisor to filter out the supervisor container from the current and target
state to let the OS handle the creation and update of the supervisor
container.

Updating and creating the supervisor container will be handled by a
future commit